Moves jest files to test folder
Updates network used in tests from Binance to Avalanche due to RPC connectivity issues
