+++ src/index.ts
@@ -1,5 +1,5 @@
 // re-export
-export { DappwrightJestConfig } from './jest/global';
+export { DappwrightJestConfig } from '../test/jest/global';
 export { getMetamask, getMetamaskWindow } from './metamask';
 export * from './setup';
 export * from './types';
